[main] img2webp 옵션을 활용해 이미지마다 lossy compression mode를 적용함

+++ src/main/java/webp/VideoCaptureThread.java
@@ -43,7 +43,7 @@ public class VideoCaptureThread extends Thread {
                             BufferedImage bufferedImage = AWTUtil.toBufferedImage(picture);
                             BufferedImage resizedImage = resize(bufferedImage, (int) (bufferedImage.getWidth() / 4.4), (int) (bufferedImage.getHeight() / 4.4));
                             ImageIO.write(resizedImage, "png",
-                                    new File("./images/" + (int)(m/4) + ".png"));
+                                    new File("./images/" + (m/4) + ".png"));
                         }
 
                     }
